--- a/python/ray/serve/benchmarks/baselines/Makefile
+++ b/python/ray/serve/benchmarks/baselines/Makefile
@@ -6,3 +6,7 @@ uvicorn:
 
 fastapi:
 	uvicorn fastapi_app:app --no-access-log --workers 1
+
+bench:
+	wrk -c 100 -t 10 -d 10s http://127.0.0.1:8000
+