--- a/python/ray/serve/benchmarks/baselines/fastapi_app.py
+++ b/python/ray/serve/benchmarks/baselines/fastapi_app.py
@@ -1,0 +1,8 @@
+from fastapi import FastAPI
+
+app = FastAPI()
+
+
+@app.get("/")
+async def read_root():
+    return "Hello world"