--- a/python/ray/serve/benchmarks/baselines/uvicorn_app.py
+++ b/python/ray/serve/benchmarks/baselines/uvicorn_app.py
@@ -1,0 +1,13 @@
+async def app(scope, receive, send):
+    assert scope['type'] == 'http'
+    await send({
+        'type': 'http.response.start',
+        'status': 200,
+        'headers': [
+            [b'content-type', b'text/plain'],
+        ]
+    })
+    await send({
+        'type': 'http.response.body',
+        'body': b'Hello, world!',
+    })