--- a/python/ray/serve/http_proxy.py
+++ b/python/ray/serve/http_proxy.py
@@ -1,5 +1,4 @@
 import asyncio
-import socket
 
 import uvicorn