<!-- Thank you for your contribution! Please review https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before opening a pull request. -->

## Why are these changes needed?
This PR
- remove some dead code in HTTP server and simplify the logic.
- add a few benchmark scripts which is helpful comparing performance improvement (result shown below)
- use uvloop if possible, since it's faster than stdlib asyncio when dealing with network IO.

## Performance

*Baselines*:

Uvicorn:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.47ms  658.62us   7.45ms   66.07%
Requests/sec:  40453.06
```

FastAPI:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     8.76ms    2.10ms  20.39ms   69.87%
Requests/sec:  11345.44
```

*master*:
pinging /noop endpoint, which returns hello world
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   145.89ms   17.37ms 195.63ms   84.16%
Requests/sec:    677.24
```

pinging / endpoint, which returns 404, doesn't go through router
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.02ms   20.82ms 242.31ms   97.81%
Requests/sec:  15807.44
```

*This PR*:
pinging /noop
```
 Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   132.56ms   15.33ms 191.02ms   84.19%
Requests/sec:    744.57
```

pinging / 404
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.79ms    6.73ms 106.18ms   97.41%
Requests/sec:  23680.14
```


<!-- Please give a short summary of the change and the problem this solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for https://docs.ray.io/en/latest/.
- [ ] I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failure rates at https://ray-travis-tracker.herokuapp.com/.
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested (please justify below)